+++ hpc/metax/samples/cuda_adapter.h
@@ -1,0 +1,1 @@
+../../nvidia/cuda/cuda_adapter.h