+++ hpc/nvidia/cuda/samples/blas.cu
@@ -21,7 +21,7 @@
 # include <hcblas/hcblas.h>
 # include <hcblas/hcblasLt.h>
 # include "hpcc_helpers.h"
-# include "cuda2hccl.h"
+# include "cuda_adapter.h"
 # define NAME	"MetaX BLAS"
 #else
 # include <cuda_runtime.h>