+++ hpc/nvidia/cuda/samples/cudaEvent.cu
@@ -9,7 +9,7 @@
 #if defined(HAVE_HCCL)
 # include <hc_runtime.h>
 # include "hpcc_helpers.h"
-# include "cuda2hccl.h"
+# include "cuda_adapter.h"
 #else
 # include <cuda_runtime.h>
 # include "cuda_helpers.h"